finish KernelFuns for Periodic Intervals
Works great for functions on periodic intervals.

Not sure how this reasoning can work with functions on Circles. For
instance, 1/z = 1/(y-x) doesn’t have a finite expansion in terms of x^i
y^j for i,j \in \mathbb{Z}.

But I’m also not sure if it will be necessary.