tidy up
Typed GreensFun{K<:BivariateFun}

but it extracts other GreensFun’s so that it ultimately only contains
other BivariateFuns.

Typing GreensFun’s will be relevant for woodbury_solve.

The tests pass with some tests commented out. The failures have to do
with new SumSpaces and integration of CosSpace() (ends up being a
SumSpace of Chebyshev and CosSpace)